+++ R/Aaaa.R
@@ -3,6 +3,7 @@
 ##' @importFrom gridExtra grid.arrange arrangeGrob
 ##' @importFrom grid grid.draw gpar grid.newpage
 ##' @importFrom grid textGrob gList
+##' @importFrom assertthat assert_that
 ##'
 NULL