+++ NEWS.md
@@ -2,6 +2,9 @@
 - Added ncol as a formal argument; when ncol > 1, we set arrange to TRUE
 - Added defaults for script and stem for mrggdraw so that 
 plots could be drawn without adding this information
+- Added `mrggpage` function as a wrapper around `gridExtra::arrangeGrob`
+- Added `mrggsave_list`, which takes in a list of objects that are ready 
+to label and to save without arrangement
 
 # 0.1.0
 - Initial validation